requirements.txt
MarkupSafe==2.0.1 
Send2Trash==1.5.0 
absl-py==0.12.0 
altair==4.1.0 
argon2-cffi==20.1.0 
astor==0.8.1 
async-generator==1.10 
attrs==21.2.0 
backcall==0.2.0 
base58==2.0.0 
bleach==3.3.0 
blinker==1.4 
boto3==1.17.80 
botocore==1.20.80 
cachetools==4.2.2 
certifi==2020.12.5 
cffi==1.14.5 
chardet==4.0.0 
click==8.0.1 
cycler==0.10.0 
decorator==5.0.9 
defusedxml==0.7.1 
entrypoints==0.3 
enum-compat==0.0.3 
idna==2.10 
importlib-metadata==4.0.1 
ipykernel==5.5.5 
ipython==7.23.1 
ipython-genutils==0.2.0 
ipywidgets==7.6.3 
jedi==0.18.0 
jinja2==3.0.1 
jmespath==0.10.0 
joblib==1.0.1 
jsonschema==3.2.0 
jupyter-client==6.1.12 
jupyter-core==4.7.1 
jupyterlab-pygments==0.1.2 
jupyterlab-widgets==1.0.0 
kiwisolver==1.3.1 
matplotlib>=1.4.3
mediapipe==0.8.4.2 
mistune==0.8.4 
nbclient==0.5.3 
nbconvert==6.0.7 
nbformat==5.1.3 
nest-asyncio==1.5.1 
notebook==6.1.6 
numpy>=1.9.2
opencv-contrib-python==4.5.2.52 
opencv-python-headless==4.5.2.52 
packaging==20.9 
pandas>=0.19
pandocfilters==1.4.3 
parso==0.8.2 
pexpect==4.8.0 
pickleshare==0.7.5 
pillow==8.2.0 
prometheus-client==0.10.1 
prompt-toolkit==3.0.18 
protobuf==3.17.1 
ptyprocess==0.7.0 
pycparser==2.20 
pydeck==0.6.2 
pygments==2.9.0 
pyparsing==2.4.7 
pyrsistent==0.17.3 
python-dateutil==2.8.1 
pytz==2021.1 
pyzmq==22.0.3 
requests==2.25.1 
s3transfer==0.4.2 
scikit-learn==0.24.2 
scipy==1.6.3 
setuptools==57.0.0 
six==1.16.0 
streamlit==0.61.0 
terminado==0.10.0 
testpath==0.5.0 
threadpoolctl==2.1.0 
toml==0.10.2 
toolz==0.11.1 
tornado==5.1.1 
traitlets==5.0.5 
typing-extensions==3.10.0.0 
tzlocal==2.1 
urllib3==1.26.4 
validators==0.18.2 
watchdog==2.1.2 
wcwidth==0.2.5 
webencodings==0.5.1 
wheel==0.36.2 
widgetsnbextension==3.5.1 
zipp==3.4.1

